requirements.txt
kafka-python==2.2.15
pika==1.3.2
matplotlib
numpy